Struct piston_window::deform::DeformGrid

pub struct DeformGrid {
    pub cols: usize,
    pub rows: usize,
    pub rect: [f64; 4],
    pub vertices: Vec<[f64; 2], Global>,
    pub indices: Vec<usize, Global>,
    pub texture_coords: Vec<[f32; 2], Global>,
    pub ps: Vec<[f64; 2], Global>,
    pub qs: Vec<[f64; 2], Global>,
    pub wis: Vec<f64, Global>,
}

Represents a deformed grid.

Fields
cols: usize

The number of columns in the grid.

rows: usize

The number of rows in the grid.

rect: [f64; 4]

The grid undeformed, which is a plain rectangle.

vertices: Vec<[f64; 2], Global>

The vertices, deformed.

indices: Vec<usize, Global>

The triangle indices.

texture_coords: Vec<[f32; 2], Global>

The texture coordinates.

ps: Vec<[f64; 2], Global>

Initial position of control points.

qs: Vec<[f64; 2], Global>

The current position of control points.

wis: Vec<f64, Global>

A weight computation buffer, one for each control point.

Implementations

impl DeformGrid

pub fn new(rect: [f64; 4], cols: usize, rows: usize) -> DeformGrid

Creates a new DeformGrid.

pub fn set_current(&mut self, i: usize, pos: [f64; 2])

Sets current control position.

pub fn set_original(&mut self, i: usize, pos: [f64; 2])

Sets original control position.

pub fn reset_control_points(&mut self)

Removes all control points.

pub fn reset_vertices_and_texture_coords(&mut self)

Sets vertices and texture coords back to default.

pub fn hit(&self, pos: [f64; 2]) -> Option<[f64; 2]>

Finds original coordinate. If the deformed grid is overlapping itself, multiple hits might occur. Returns the first hit it finds.

pub fn draw_image<G>(
&self,
texture: &<G as Graphics>::Texture,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
)where
G: Graphics,

Draws deformed image using default method.

pub fn draw_image_tri<G>(
&self,
texture: &<G as Graphics>::Texture,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
)where
G: Graphics,

Draws deformed image using triangulation.

pub fn add_control_point(&mut self, pos: [f64; 2])

Adds a control point, in original coordinates.

pub fn draw_vertical_lines<G>(
&self,
line: &Line,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
)where
G: Graphics,

Draw vertical grid lines.

pub fn draw_horizontal_lines<G>(
&self,
line: &Line,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
)where
G: Graphics,

Draw horizontal grid lines.

pub fn update(&mut self)

Updates the grid, by deforming the vertices.