Struct piston_window::Rectangle

pub struct Rectangle {
    pub color: [f32; 4],
    pub shape: Shape,
    pub border: Option<Border>,
}

A filled rectangle

Fields
color: [f32; 4]

The rectangle color

shape: Shape

The roundness of the rectangle

border: Option<Border>

The border

Implementations

impl Rectangle

pub fn new(color: [f32; 4]) -> Rectangle

Creates a new rectangle.

pub fn new_round(color: [f32; 4], round_radius: f64) -> Rectangle

Creates a new rectangle with rounded corners.

pub fn new_border(color: [f32; 4], radius: f64) -> Rectangle

Creates a new rectangle border.

pub fn new_round_border(
color: [f32; 4],
round_radius: f64,
border_radius: f64
) -> Rectangle

Creates a new rectangle border with rounded corners.

pub fn color(self, value: [f32; 4]) -> Rectangle

Sets color.

pub fn shape(self, value: Shape) -> Rectangle

Sets shape of the corners.

pub fn border(self, value: Border) -> Rectangle

Sets border properties.

pub fn maybe_border(self, value: Option<Border>) -> Rectangle

Sets optional border.

pub fn draw_from_to<P, G>(
&self,
from: P,
to: P,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
) where
P: Into<[f64; 2]>,
G: Graphics,

Draws the rectangle by corners using the default method.

pub fn draw<R, G>(
&self,
rectangle: R,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
) where
R: Into<[f64; 4]>,
G: Graphics,

Draws the rectangle using the default method.

rectangle defines the rectangle's location and dimensions, draw_state draw state, draw_state::Default::default() can be used as a default, transform is the transformation matrix, g is the Graphics implementation, that is used to actually draw the rectangle.s

pub fn draw_tri<R, G>(
&self,
rectangle: R,
draw_state: &DrawState,
transform: [[f64; 3]; 2],
g: &mut G
) where
R: Into<[f64; 4]>,
G: Graphics,

Draws the rectangle using triangulation.

This is the default implementation of draw() that will be used if G does not redefine Graphics::rectangle().