Trait piston_window::SourceRectangled

pub trait SourceRectangled {
    fn src_rect(self, x: f64, y: f64, w: f64, h: f64) -> Self;
    fn src_rel(self, x: f64, y: f64) -> Self;
    fn src_flip_h(self) -> Self;
    fn src_flip_v(self) -> Self;
    fn src_flip_hv(self) -> Self;
}

Should be implemented by contexts that have source rectangle information.

Required Methods

fn src_rect(self, x: f64, y: f64, w: f64, h: f64) -> Self

Adds a source rectangle.

fn src_rel(self, x: f64, y: f64) -> Self

Moves to a relative source rectangle using the current source rectangle as tile.

fn src_flip_h(self) -> Self

Flips the source rectangle horizontally.

fn src_flip_v(self) -> Self

Flips the source rectangle vertically.

fn src_flip_hv(self) -> Self

Flips the source rectangle horizontally and vertically.