Struct piston_window::context::Context

pub struct Context {
    pub viewport: Option<Viewport>,
    pub view: [[f64; 3]; 2],
    pub transform: [[f64; 3]; 2],
    pub draw_state: DrawState,
}

Drawing 2d context.

Fields
viewport: Option<Viewport>

Viewport information.

view: [[f64; 3]; 2]

View transformation.

transform: [[f64; 3]; 2]

Current transformation.

draw_state: DrawState

Current draw state settings.

Methods

impl Context

pub fn new() -> Context

Creates a new drawing context.

pub fn new_viewport(viewport: Viewport) -> Context

Creates a new context with absolute transform in point coordinates.

This function assumes the default coordinate system being centered with x axis pointing to the right and y axis pointing up.

Returns a drawing context with origin in the upper left corner and x axis pointing to the right and y axis pointing down.

pub fn new_abs(w: f64, h: f64) -> Context

Creates a new drawing context in absolute coordinates.

This function assumes the default coordinate system being centered with x axis pointing to the right and y axis pointing up.

Returns a drawing context with origin in the upper left corner and x axis pointing to the right and y axis pointing down.

pub fn view(self) -> Context

Moves the current transform to the view coordinate system.

This is usually [0.0, 0.0] in the upper left corner with the x axis pointing to the right and the y axis pointing down.

pub fn reset(self) -> Context

Moves the current transform to the default coordinate system.

This is usually [0.0, 0.0] in the center with the x axis pointing to the right and the y axis pointing up.

pub fn store_view(self) -> Context

Stores the current transform as new view.

pub fn get_view_size(&self) -> [f64; 2]

Computes the current view size.