Struct piston_window::PistonWindow

pub struct PistonWindow<W: Window = GlutinWindow> {
    pub window: W,
    pub encoder: GfxEncoder,
    pub device: Device,
    pub output_color: RenderTargetView<Resources, Srgba8>,
    pub output_stencil: DepthStencilView<Resources, DepthStencil>,
    pub g2d: Gfx2d<Resources>,
    pub events: Events,
    pub factory: Factory,
}

Contains everything required for controlling window, graphics, event loop.

Fields
window: W

The window.

encoder: GfxEncoder

GFX encoder.

device: Device

GFX device.

output_color: RenderTargetView<Resources, Srgba8>

Output frame buffer.

output_stencil: DepthStencilView<Resources, DepthStencil>

Output stencil buffer.

g2d: Gfx2d<Resources>

Gfx2d.

events: Events

Event loop state.

factory: Factory

The factory that was created along with the device.

Methods

impl<W> PistonWindow<W> where
W: Window,

pub fn new(opengl: OpenGL, samples: u8, window: W) -> Self where
W: OpenGLWindow,

Creates a new piston window.

pub fn draw_2d<E, F, U>(&mut self, e: &E, f: F) -> Option<U> where
W: OpenGLWindow,
E: GenericEvent,
F: FnOnce(Context, &mut G2d) -> U,

Renders 2D graphics.

Calls the closure on render events. There is no need to filter events manually, and there is no overhead.

pub fn draw_3d<E, F, U>(&mut self, e: &E, f: F) -> Option<U> where
W: OpenGLWindow,
E: GenericEvent,
F: FnOnce(&mut Self) -> U,

Renders 3D graphics.

Calls the closure on render events. There is no need to filter events manually, and there is no overhead.

pub fn next(&mut self) -> Option<Event>

Returns next event. Cleans up after rendering and resizes frame buffers.

pub fn event<E: GenericEvent>(&mut self, event: &E)

Let window handle new event. Cleans up after rendering and resizes frame buffers.