Struct piston_window::PistonWindow

pub struct PistonWindow<W: Window = GlutinWindow, T = ()> {
    pub window: Rc<RefCell<W>>,
    pub stream: Rc<RefCell<GfxStream>>,
    pub device: Rc<RefCell<Device>>,
    pub g2d: Rc<RefCell<Gfx2d<Resources>>>,
    pub events: Rc<RefCell<WindowEvents<W, Event<W::Event>>>>,
    pub event: Option<Event<W::Event>>,
    pub app: Rc<RefCell<T>>,
    pub factory: Rc<RefCell<Factory>>,
}

Contains everything required for controlling window, graphics, event loop.

Fields

window

The window.

stream

GFX stream.

device

GFX device.

g2d

Gfx2d.

events

The event loop.

event

The event.

app

Application structure.

factory

The factory that was created along with the device.

Methods

impl<W, T> PistonWindow<W, T> where W: Window, W::Event: GenericEvent

fn new(window: Rc<RefCell<W>>, app: Rc<RefCell<T>>) -> Self where W: OpenGLWindow

Creates a new piston object.

fn app<U>(self, app: Rc<RefCell<U>>) -> PistonWindow<W, U>

Changes application structure.

fn draw_2d<F>(&self, f: F) where F: FnOnce(Context, &mut GfxGraphics<Resources, CommandBuffer, Output>)

Renders 2D graphics.

fn draw_3d<F>(&self, f: F) where F: FnOnce(&mut GfxStream)

Renders 3D graphics.